BackEnd/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10  # PostgreSQL driver
python-dotenv==1.0.1     # For loading .env (optional, but handy)
argon2-cffi